Split `SWT_NO_EXIT_TESTS` into `SWT_NO_EXIT_TESTS` and `SWT_NO_PROCESS_SPAWNING`.
This PR separates out our process-spawning code to be guarded by
`SWT_NO_PROCESS_SPAWNING` instead of `SWT_NO_EXIT_TESTS`. We do this so that we
can potentially use process spawning on platforms where exit tests are not
supported for some other reason (such as the iOS/Android sandboxes) but process
spawning is still internally possible.

There are a few use cases we have for spawning processes that don't involve exit
tests:

- Calling out to `tar` to compress attachments (see #714)
- Running non-Swift scripts in their interpreters (see #478)
- Multi-process parallelism (the XCTest model)

I took the opportunity to clean up WaitFor.swift a bit and rearrange code so
that the "new platform, dunno who lives here" case should compile (although not
function) out-of-the-box.