Add a return type for exit tests. (#762)
This PR adds a new structure, `ExitTest.Result`, that is returned from a
call to `#expect(exitsWith:)` or `#require(exitsWith:)`. It currently
contains the actual exit condition of the test (which should equal the
expected exit condition on success) but in the future we can add
additional properties such as the standard output and standard error
streams.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.