Update documentation to explain misconfigurations of `SWT_` conditions.
This PR was supposed to be part of #769, but I must have blown away a local
commit because it didn't make it into that PR.

This PR updates Porting.md to explain when a developer might encounter an error
due to mismatched `SWT_` conditions. It ensures that exit tests require process
spawning to be enabled/implemented by adding a dependency between
`SWT_NO_PROCESS_SPAWNING` and `SWT_NO_EXIT_TESTS`.

This PR also tweaks which bits of `ExitTest` are exposed on platforms that do
not support exit tests at all. The type needs to be exposed so that
`ExitTest.Result` is exposed so that we can correctly expose
`#expect(exitsWith:)` and `#require(exitsWith:)` while marking them unavailable,
as opposed to hiding the macros and not giving a developer clear diagnostics.