Add support for collecting `stdout` and `stderr` in an exit test.
This PR adds `stdout` and `stderr` capture to exit tests.

Gathering these streams can be very expensive, so they are _opt-in_. There is a
new argument to `#expect(exitsWith:)`, `observedValues`, to which you can pass
one or more key paths on `ExitTest.Result` to indicate which bits you want:

```swift
 let result = await #expect(exitsWith: .failure, observing: [\.standardOutputContent]) {
   print("Goodbye, world!")
   fatalError()
 }
 #expect(result.standardOutputContent.contains(UInt8(ascii: "G")))
```

Some changes to pipes were needed to satisfy the compiler's concerns about
`Pipe` being a move-only type that contains other move-only types—now it's just
a function that returns two `FileHandle` instances. This is simpler anyway.

It is possible some callers will want `stdout` and `stderr` to be set, but to
something other than a pipe (such as a file on disk or a PTY.) While these are
interesting cases, they are beyond the scope of this PR.