Avoid using `NTSTATUS` near `exit()`.
On Windows, `NTSTATUS` is a typedef of `long`. On 32-bit Windows, `long` is
imported into Swift as `Int` but on 64-bit Windows,it's imported as `Int32`.
These types are layout-equivalent but not implicitly convertible, so we get
errors on 32-bit Windows when using an `NTSTATUS` (i.e. `Int`) where a `CInt`
(i.e. `Int32`) is expected.

This PR changes the types of the constants introduced in #766 to be explicitly
`CInt` so that they are usable on both flavours of Windows.