Attachments! (#770)
This PR introduces a new experimental feature, attachments. With this
feature, you can "attach" values that conform to a new `Test.Attachable`
protocol to a test. With the right command-line incantation (TBD), Swift
Testing will automatically write attachments to disk for you.

> [!NOTE]
> This PR does not teach Xcode or VS Code how to handle attachments
produced by
> Swift Testing, nor does it add the necessary command-line arguments to
the
> `swift test` command-line tool.

This PR is one of a series that I'll be posting to build out this
feature.

As always, keep in mind that symbols marked `@_spi(Experimental)` are
subject to change or removal without notice.

Resolves #714.
Resolves rdar://88648735.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.