Specify the module to find the ABI entry point function in (in our own tests).
The order in which Xcode loads the Apple-vendored copy of Swift Testing vs. a
developer-supplied XCTest bundle is unspecified. Our tests assume that the
"current" copy of Swift Testing (that is, the copy they linked against) is the
first one loaded, which is not always true when running tests in Xcode. The
modified tests end up finding the wrong entry point functions when called within
Xcode if the load order has changed.

This PR explicitly specifies that we are looking for the entry point functions
in the module the tests are linked against.

Resolves rdar://139140302.