Eagerly move attachable values to the heap.
This PR modifies `Test.Attachment.init(...)` so that, if the attachable value
passed to it is of a large or non-bitwise-copyable value type, we store it in
a heap-allocated container object.

Why do we want to do this? Attachable values have arbitrary types we don't
really know anything about, and may be very expensive to copy if they are large
or if they contain many references to other objects. Since attachments are
sendable and need to pass through many layers in the system, the cost of making
copies for these attachable values can add up quickly. So, instead of copying
them repeatedly, if we store them in a refcounted container and pass _that_
value around, we only infer a single copy cost.

There are four modes for attachable value storage after this PR:

| Type | Where |
|-|-|
| Refcounted, sendable objects | On the heap |
| Sendable, small, POD values | Inline in an existential |
| Sendable, large-or-non-POD values | In a heap-allocated box |
| Non-sendable or move-only values | On the heap, eagerly serialized |