Provide an opt-out conditional for lazy attachment encoding.
This PR provides a compile-time conditional, `SWT_NO_LAZY_ATTACHMENTS`, which
when set causes Swift Testing's new experimental attachments feature to always
eagerly encode/serialize/whatever attachments even if they're copyable and
sendable.

We anticipate using this functionality in Embedded Swift in the future. Note
that support for Embedded Swift is not (yet!) a planned feature.