Add a CoreGraphics cross-import overlay with support for attaching `CGImage`s. (#827)
This PR adds a new cross-import overlay target with Apple's Core
Graphics framework that allows attaching a `CGImage` as an attachment in
an arbitrary image format (PNG, JPEG, etc.)

Because `CGImage` is imported into Swift as a non-final class, it cannot
conform directly to `Attachable`, so an `AttachableContainer` type acts
as a proxy. This type is not meant to be used directly, so its name is
underscored. Initializers on `Attachment` are provided so that this
abstraction is almost entirely transparent to test authors.

A new protocol, `AttachableAsCGImage`, is introduced that abstracts away
the relationship between the attached image and Core Graphics; in the
future, I intend to make additional image types like `NSImage` and
`UIImage` conform to this protocol too.

Example usage:

```swift
let sparklyDiamonds: CGImage = ...
let attachment = Attachment(image, named: "sparkly-diamonds", as: .tiff, encodingQuality: 0.75)
...
attachment.attach()
```

The code in this PR is, by definition, specific to Apple's platforms. In
the future, I'd be interested in adding Windows/Linux equivalents
(`HBITMAP`? Whatever Gnome/KDE/Qt use?) but that's beyond the scope of
this PR.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.