Add a `preferredName(for:basedOn:)` member to `Attachable` to allow customizing filenames. (#854)
This PR introduces a new optional member of `Attachable`,
`preferredName(for:basedOn:)`, that we use when writing the
corresponding attachment to disk/test reports/etc. in order to allow
attachable types to customize the name independently of what the user
specifies.

For example:

```swift
let a = Attachment(x)
let b = Attachment(y, named: "hello")
```

In both the attachments created above, the file name is incompletely
specified. `a` has a default name, and both `a` and `b` have no path
extension (which is important for the OS to correctly recognize the
produced file's type.) By adding this new function to `Attachable`, we
give `x` and `y` the opportunity to say "this is JPEG data" or "this is
plain text" (and so forth.)

The new function is implemented by `_AttachableImageContainer`. I've
also created `_AttachableURLContainer` to represent files mapped from
disk for attachment (instead of directly passing them around as `Data`.)

Third-party conforming types will generally want to use Foundation's
`NSString` or `URL` API to append path extensions (etc.)

> [!NOTE]
> Attachments remain experimental.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.