Add example for how to test any error is thrown in the error testing documentation article (#853)
I was stumped earlier when trying to find the details for how to test
that any error is thrown. This detail _is_ in the migrating from XCTest
content, but I landed on this page, and it seemed an odd miss,
especially given the history of ignoring the kind of error in many
cases. (I found my solution in [a post by Jonathan the
forums](https://forums.swift.org/t/swift-testing-whats-the-recommend-approach-to-test-throwing-function/70806/2).)

While it is duplicated content, this would have helped me - so I'm
proposing it for a documentation update. Happy to edit as y'all see fit.

---------

Co-authored-by: Stuart Montgomery <smontgomery@apple.com>