Change default behavior of exit tests to always succeed.
Exit tests simulate calling `main()` (or some similar thing), but if you don't explicitly
exit from them, they force a failure. After much mulling and some discussion with
colleagues a while back, we should change the behavior so that, if a test doesn't
otherwise terminate, it acts as if `main()` returned naturally—that is, by exiting with
`EXIT_SUCCESS` rather than by forcing `EXIT_FAILURE`. This behavior is more consistent
with the feature's theory of operation.