Don't use `@_semantics`, use our own attribute instead.
This PR replaces our use of `@_semantics` with a custom attribute macro. `@_semantics` is
reserved for use by the standard library and runtime and, while useful, can be emulated on
our side of the module barrier without much difficulty.

Note there are no unit tests for this macro: it doesn't actually _do_ anything on its own!