Mark `#expect(_:throws:)` and `#require(_:throws:)` as to-be-deprecated.
These symbols should go through a transition period where they are documented as
to-be-deprecated rather than being deprecated outright. My mistake.

With this change applied, Xcode treats these macros as deprecated (deemphasizing them in
autocomplete, for instance) and DocC marks them as deprecated in an unspecified Swift
version, but no warning is emitted if they are used. This is consistent with API marked
`API_TO_BE_DEPRECATED` or `deprecated: 100000.0` in Apple's SDKs.

> [!NOTE] There is a bug in the DocC compiler that emits a diagnostic of the form:
>
> > ⚠️ 'require(_:sourceLocation:performing:throws:)' isn't unconditionally deprecated
>
> This issue is being tracked already with rdar://141785948.

Resolves #873.