Add an attribute that emits a warning.
This PR adds an attribute (an overload of `@__testing()`) that emits a warning.
This allows us to emit compile-time warnings from contexts where only attributes
are semantically valid.

We need this in particular for test content records because their section info
is comprised of a big `IfConfigDecl` node and platforms that aren't covered
(i.e. the `#else` clause) don't have a way to emit a diagnostic that says "we
need to fix this platform's Swift Testing port!" like we can do in other
contexts with `#warning()`:

```swift
@_section("__DATA_CONST,__swift5_tests")
@_section("swift5_tests")
@_section(".sw5test$B")
// ⚠️ #warning isn't valid here, what do we do!?
@_used
private static let record: __TestContentRecord = (...)
```