Remove SPIAwareTrait and adopt TestScoping in its place (#901)
This is a follow-on from #733 which introduced the Test Scoping Traits
feature. As that proposal indicated, the `SPIAwareTrait` SPI protocol is
no longer necessary, so this PR removes it and adopts `TestScoping` in
its place.

### Motivation:

Adopting Test Scoping Traits for the purpose which `SPIAwareTrait`
previously served simplifies a lot of runner and planning logic.

### Modifications:

- Delete `SPIAwareTrait` and related SPI.
- Adopt `TestScoping` in `ParallelizationTrait` instead.
- Changed `ParallelizationTrait` to _not_ always have the value `true`
for `isRecursive`. Test Scoping Traits makes this no longer necessary:
the invocation of children of any suites which have `.serialized` is
scoped to have parallelization disabled.
- Removed tracking of `isParallelizationEnabled` on the
`Runner.Plan.Action.RunOptions` type. For now, I kept the `RunOptions`
struct, since it could be useful in the future and it's SPI.
- Removed obsolete logic in `Runner` for propagating ancestor steps, and
refactored slightly to make many things `static`. This is to help ensure
certain functions don't access the `self.configuration` property when
instead they should be accessing `Configuration.current`.

### Result:

No functional change, but implementation is simplified.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.