Don't capture arguments to `#expect(exitsWith:)` during macro expansion. (#912)
Other than the body closure, the arguments to `#expect(exitsWith:)` are
not guaranteed to be literals or otherwise context-free, but we capture
them during macro expansion. This can result in obscure errors if the
developer writes an exit test with an argument that is computed:

```swift
let foo = ...
await #expect(exitsWith: self.exitCondition(for: foo)) {
  // 🛑 error: closure captures 'foo' before it is declared
  // 🛑 error: enum declaration cannot close over value 'self' defined in outer scope
  // (and other possible errors)
  ...
}
```

This PR removes the macro's compile-time dependency on its
`exitCondition` and `sourceLocation` arguments and introduces an `ID`
type to identify exit tests (instead of using their source locations as
unique IDs.) The `ID` type is meant to be a UUID, but for the moment
I've just strung together two 64-bit integers instead to avoid a
dependency on Foundation or platform API.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.