Refactor test content discovery to produce a sequence.
This PR does away with `enumerateTestContent {}` and replaces it with
`discover()`, a function that returns a sequence of `TestContentRecord`
instances that the caller can then inspect. This simplifies the logic in callers
by letting them use sequence operations/monads like `map()`, and simplifies the
logic in the implementation by doing away with intermediate tuples and allowing
it to just map section bounds to sequences of `TestContentRecord` values.

Evaluation of the accessor functions for test content records is now lazier,
allowing us to potentially leverage the `context` field of a test content record
to avoid calling accessors for records we know won't match due to the content of
that field. (This functionality isn't currently needed, but could be useful for
future test content types.)