Revert #916 and apply a slightly different change.
Thank you to @tshortli for the patch! We don't actually _need_ this property to
be unavailable on WASI—it was just meant as a convenience. But since we're not
using the property for anything at this time (a future PR might try to
productize `TestContentRecord`), it's overkill.

So instead I'll just leave the property available, and document it's always
`nil` on statically-linked platforms.