Revert #916 and apply a slightly different change. (#917)
Thank you to @tshortli for the patch! We don't actually _need_ this
property to be unavailable on WASI—it was just meant as a convenience.
But since we're not using the property for anything at this time (a
future PR might try to productize `TestContentRecord`), it's overkill.

So instead I'll just leave the property available, and document it's
always `nil` on statically-linked platforms.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.