Make all test content types directly conform to `TestContent`.
This PR eliminates the `TestContentAccessorResult` associated type from the
(currently internal, potentially eventually API) `TestContent` protocol. This
associated type needed to be `~Copyable` so `ExitTest` could be used with it,
but that appears to pose some _problems_ for the compiler (rdar://143049814&143080508).

Instead, we remove the associated type and just say "the test content record is
the type that conforms to `TestContent`". `ExitTest` is happy with this, but
`Test`'s produced type is a non-nominal function type, so we wrap that function
in a small private type with identical layout and have that type conform.

The ultimate purpose of this PR is to get us a bit closer to turning
`TestContent` into a public or tools-SPI protocol that other components can use
for test discovery.