[SWT-NNNN] Attachments
Test authors frequently need to include out-of-band data with tests that can be
used to diagnose issues when a test fails. This proposal introduces a new API
called "attachments" (analogous to the same-named feature in XCTest) as well as
the infrastructure necessary to create new attachments and handle them in tools
like VS Code.

Read the full proposal [here]().