Add a `type` argument to the ABI accessor function signature. (#939)
We access the test content record section directly instead of using some
(non-existent) type-safe Swift API. On some platforms (Darwin in
particular), it is possible for two copies of the testing library to be
loaded at runtime, and for them to have incompatible definitions of
types like `Test`. That means one library's `@Test` macro could produce
a test content record that is then read by the other library's discovery
logic, resulting in a stack smash or worse.

We can resolve this issue by adding a `type` argument to the accessor
function we define for test content records; the body of the accessor
can then compare the value of this argument against the expected Swift
type of its result and, if they don't match, bail early.

The new argument is defined as a pointer _to_ a Swift type rather than
as a Swift type directly because `@convention(c)` functions cannot
directly reference Swift types. The value of the new argument can be
safely ignored if the type of the test content record's value is and
always has been `@frozen`.

### Checklist:

- [x] Code and documentation should follow the style of the [Style
Guide](https://github.com/apple/swift-testing/blob/main/Documentation/StyleGuide.md).
- [x] If public symbols are renamed or modified, DocC references should
be updated.