Use `os_unfair_lock` on Darwin when available.
This PR refactors the `Locked` type to be generic over the type of lock it uses,
allowing us to adopt `os_unfair_lock` in most places, but also allowing us to
use other lock types as needed (i.e. `pthread_mutex_t` in WaitFor.swift if
libdispatch is unavailable.)

Under `-O`, locking and unlocking is fully inlined, while initialization (which
is relatively rare) is partially, but not completely, inlined.