Add a (internal-only) `CustomIssueRepresentable` protocol.
This PR adds a (for now internal-only) protocol that lets us transform arbitrary
errors to arbitrary issues (rather than always emitting them as `.errorCaught`).
This protocol replaces the special-casing we were doing for `SystemError` and
`APIMisuseError`.

This change also causes us to emit an API misuse issue if a developer passes an
error of type `ExpectationFailedError` to `Issue.record()` as we can reasonably
assume that such an error was already recorded correctly as `.expectationFailed`.