Suppress warning about `#require(nonOptional)` in some cases.
When using `try #require()` to unwrap an optional value, we emit a compile-time
warning if the value is not actually optional. However, there is a bug in the
type checker (https://github.com/swiftlang/swift/issues/79202) that triggers a
false positive when downcasting an object (of `class` type), e.g.:

```swift
class Animal {}
class Duck: Animal {}
let beast: Animal = Duck()
let definitelyADuck = try #require(beast as? Duck)
// ⚠️ '#require(_:_:)' is redundant because 'beast as? Duck' never equals 'nil'
```

This change suppresses the warning we emit if the expression contains certain
syntax tokens (namely `?`, `nil`, or `Optional`) on the assumption that their
presence means the test author is expecting an optional value and we've hit a
false positive.