Plumb the schema version through the JSON ABI layers.
This PR plumbs the schema version (0 or 1) through the Swift types and functions
that produce our JSON output, allowing them to change behaviour based on which
version is in use.

As a proof-of-concept as much as anything else, this PR also adds an
(unsupported) `_tags` field to the JSON output for a test record, but only with
ABI version 1.